% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse.wav.fingerprint}
\alias{analyse.wav.fingerprint}
\title{Analyse WaveData for Fingerprints}
\usage{
analyse.wav.fingerprint(wavdata, onsets, fp.params = list(),
  stoptime = NULL, duration = NULL, quiet = TRUE)
}
\arguments{
\item{wavdata}{WaveData object to be analysed}

\item{onsets}{List of onset objects, from which the start and stop times 
for the extraction should be taken}

\item{fp.params}{Parameters passed to \code{\link{fingerprint}}}

\item{stoptime}{Stoptime to filter onsets in fingerprinting}

\item{duration}{Explicit duration of the part to be fingerprinted}

\item{quiet}{If set to true, the function will output the current filename to be analysed}
}
\description{
Analyse WaveData for Fingerprints
}
\details{
Fingerprints are extracted from the detected onset area with the largest energy. If a 
stoptime is given onsets before the stoptime are ignored. End times can be overridden by
providing an explicit duration in the duration parameter.
}

