% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse.wav.onsets}
\alias{analyse.wav.onsets}
\title{Analyse WaveData for Onset Times}
\usage{
analyse.wav.onsets(wavdata, onset.params = list(), energy.params = list(),
  quiet = TRUE)
}
\arguments{
\item{wavdata}{WaveData object to be analysed}

\item{onset.params}{Parameters passed to \code{\link{onsets}}}

\item{energy.params}{Additional Parameters passed to \code{\link{energyDensity}}.}

\item{quiet}{If set to true, the function will output the current filename to be analysed}
}
\description{
This function mainly serves as a wrapper for calling the onset detection
function with a different parameter format.
}

