% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsHelpers.R
\name{duration}
\alias{duration}
\title{Generic function to extract the duration of any kind of time series object}
\usage{
duration(x, ...)
}
\arguments{
\item{x}{The time series for which the duration is requested}

\item{...}{Extra arguments for future methods}
}
\description{
Generic function to extract the duration of any kind of time series object
}

