% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyDensity.R
\name{energyDensity.WaveData}
\alias{energyDensity.WaveData}
\title{energyDensity implementation for WaveData objects}
\usage{
\method{energyDensity}{WaveData}(ts, window.width = 10, stepsize = 5,
  window.function = signal::boxcar, ...)
}
\arguments{
\item{ts}{The object that contains the time series data}

\item{window.width}{The windowing width that is used to estimate the density (in ms). The time
series data is chopped into windows of this length, and the energy density is calculated
for each of these windows seperately. Defaults to 10ms.}

\item{stepsize}{The stepsize to be used for windowing. Windows are advanced this many ms for each
calculation. Can be used to create overlapping windows. Defaults to 5ms.}

\item{window.function}{The windowing function to be used. This argument can be used to give
different weigth to different samples within the window. Several usable windowing functions
   are defined in the signal package. Defaults to \code{\link[signal]{boxcar}}, i.e. equal weight
   for all samples.}

\item{...}{ignored}
}
\description{
energyDensity implementation for WaveData objects
}

