% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyDensity.R
\name{energyDensityAt}
\alias{energyDensityAt}
\title{Calculate the Energy Density at a Specific Position of a Standard R ts Object}
\usage{
energyDensityAt(ts, start, end, window.function = signal::hanning)
}
\arguments{
\item{ts}{The time series object.}

\item{start}{The starting position at which the energy density should be calculated.}

\item{end}{The end position until which the energy density should be calculated.}

\item{window.function}{A windowing function to weight the samples.}
}
\description{
Calculate the Energy Density at a Specific Position of a Standard R ts Object
}

