% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvOutput.R
\name{expResponses.as.table}
\alias{expResponses.as.table}
\title{Analyse All Files from a Given Directory and Write Results as Tabular Format}
\usage{
expResponses.as.table(dirname, filename, read.params = list(),
  filter = list(), onset.params = list(), energy.params = list(),
  fp.params = list(), stoptime = NULL, duration = NULL,
  include.fp = FALSE, nresponses, ...)
}
\arguments{
\item{dirname}{Name of the directory from which the files should be taken}

\item{filename}{The filename of the csv file to which the results should be written}

\item{read.params}{Parameters passed to \code{\link{read.wav}} (except filter parameters)}

\item{filter}{List of filter parameters (set to \code{NULL} to deactivate filter). 
See "Details"}

\item{onset.params}{Parameters passed to \code{\link{onsets}}}

\item{energy.params}{Additional Parameters passed to \code{\link{energyDensity}}.}

\item{fp.params}{Parameters passed to \code{\link{fingerprint}}}

\item{stoptime}{Stoptime to filter onsets in fingerprinting}

\item{duration}{Explicit duration of the part to be fingerprinted}

\item{include.fp}{Whether the full fingerprint should be included in the output
(default=\code{FALSE})}

\item{nresponses}{Number of response classes used in clustering. Set to \code{NULL} to 
deactivate cluster analysis}

\item{...}{Optional arguments passed to \code{\link[utils]{write.table}}}
}
\description{
Analyse All Files from a Given Directory and Write Results as Tabular Format
}

