% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprinting.R
\name{extract.fp.minmax}
\alias{extract.fp.minmax}
\title{Extract Time and Strength Independent Fingerprints}
\usage{
extract.fp.minmax(fs)
}
\arguments{
\item{fs}{The feature time series from which the fingerprints should be extracted}
}
\description{
Extract Time and Strength Independent Fingerprints
}
\details{
This function extracts fingerprints, which are somewhat invariant to stretching in the time
domain while keeping the order of features intact. Furthermore these fingerprints are
completely invariant to a scaling or offset of the complete feature series.

For each feature within the feature vectors, the times of maximum and minimum value are 
extracted. Since no values of the actual features are used, scaling and offset invariance
is achieved. Furthermore, the positions of minima and maxima are scaled relative to the
total length of the series. Therefore, these fingerprints are robust to stretching of
the complete time series. Local stretching of the feature series are translated to
movements in euclidean space.
}

