% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature.type}
\alias{feature.type}
\title{Return a type identifier for a feature time series}
\usage{
feature.type(fs)
}
\arguments{
\item{fs}{The time series object}
}
\description{
This function returns a type identifier for a feature
series. In most cases this should correspond to the
primary class of the feature time series.
}

