% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprinting.R
\name{fingerprint.Features}
\alias{fingerprint.Features}
\title{Extract Fingerprints from a time series of feature vectors (e.g. spectrum or MFCC data)}
\usage{
\method{fingerprint}{Features}(ts, fp.type = c("minmax", "mean"), ...)
}
\arguments{
\item{ts}{The feature time series from which the fingerprint should be extracted}

\item{fp.type}{The type of extraction to be performed (see details)}

\item{...}{Parameters for future methods}
}
\description{
Simple wrapper, which calls the specified extraction function.
}
\details{
This function calls the specified extraction function. For \code{type="minmax"} the
function \code{\link{extract.fp.minmax}} is called and for \code{type="mean"} the
\code{\link{extract.fp.mean}} function is called.

Fingerprints can be compared by (euclidean) distance to determine the similarity between
feature series given some invariants. The type of invariants is determined by the concrete
extraction method.
}

