% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprinting.R
\name{fingerprint}
\alias{fingerprint}
\title{Extract Fingerprints from any Kind of Time series.}
\usage{
fingerprint(ts, ...)
}
\arguments{
\item{ts}{The time series from which fingerprints should be extracted}

\item{...}{Parameters for future methods}
}
\description{
This provides a generic S3 method to extract fingerprints from any kind of time
series. The concrete type of possible fingerprints is determined by the type
of object.
}

