% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfccs.R
\name{MFCCs}
\alias{MFCCs}
\title{S3 Generic to calculate MFCCs (Mel Frequency Cepstral Coefficients) for a given time series}
\usage{
MFCCs(ts, window.width = 25, stepsize = 10,
  window.function = signal::hanning, lower = 300, upper = 8000,
  filterbanks = 26, preemph = 0.97, retain.coeffs = 1:13, delta = 2,
  ...)
}
\arguments{
\item{ts}{The time series object}

\item{window.width}{Window width for each MFCC vector (in ms)}

\item{stepsize}{Stepsize ms samples by which each window is advanced}

\item{window.function}{Windowing function for window extraction.}

\item{lower}{Lowest frequency to be used for MFCC calculation (default 300Hz)}

\item{upper}{Highest frequency to be used for MFCC calculation (default 8kHz)}

\item{filterbanks}{Number of MFCC filterbanks to produce (default 26)}

\item{preemph}{Strength of pre-emphasis filter applied to original signal (defaul 0.97).}

\item{retain.coeffs}{A vector of the coefficients to retain (starting a 0; default 1:13).
Set to \code{NULL} to keep all coefficients.}

\item{delta}{Number of deltas (differences between subsequent vectors) to be computed}

\item{...}{Additional parameters for future methods}
}
\description{
S3 Generic to calculate MFCCs (Mel Frequency Cepstral Coefficients) for a given time series
}
\details{
The signal is sliced into windows each of the size determined by \code{window.width}. For each window
a MFCC vector is calculated. For each
}

