% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset.R
\name{onsets}
\alias{onsets}
\title{Calculate (Physical) Onset Times from an Object containing Time Series Data}
\usage{
onsets(ts, limit = 0.1, ...)
}
\arguments{
\item{ts}{The object containing the time series data.}

\item{limit}{Parameter used for filtering silence. Energy below this level
(compared to all other energy levels in term of percentiles) is ignored.}

\item{...}{Object specific parameters}
}
\description{
This function calculates onset times for any kind of object containing time
series data. The onsets are calculated by comparing the overall energy of 
each part of the time series with the rest of the time series. Positions
where the energy is above a portion of \code{limit}, compared to the overal
energy, are assumed to contain actual data.
}

