% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset.R
\name{onsets.WaveData}
\alias{onsets.WaveData}
\title{Specialized Method for Extracting Onsets from WaveData objects}
\usage{
\method{onsets}{WaveData}(ts, limit = 0.1, limit.type = c("absolute",
  "relative"), window.width = 10, stepsize = 5, normalize = 0.9,
  window.function = signal::hanning, ...)
}
\arguments{
\item{ts}{The object containing the time series data.}

\item{limit}{Parameter used for filtering silence. Energy below this level
is ignored.}

\item{limit.type}{Type of limit, which is used to detect on- and offsets.
If set to \code{absolute} (default) then the limit value is used
directly. If set to \code{relative} then the limit is determined
as the energy in the i-th percentile, where i is given by the
limit parameter.}

\item{window.width}{The windowing width that is used to estimate the density (in ms). The time
series data is chopped into windows of this length, and the energy density is calculated
for each of these windows seperately. Defaults to 10ms.}

\item{stepsize}{The stepsize to be used for windowing. Windows are advanced this many ms for each
calculation. Can be used to create overlapping windows. Defaults to 5ms.}

\item{normalize}{Normalization level for energy density values. Maximum energy density will be set to this
level to simplify comparison between data files. Set to 0 to deactivate normalization. 
Defaults to 0.9. Maximum normalization to 1.}

\item{window.function}{The windowing function to be used. This argument can be used to give
different weigth to different samples within the window. Several usable windowing functions
   are defined in the signal package. Defaults to \code{\link[signal]{boxcar}}, i.e. equal weight
   for all samples.}

\item{...}{ignored}
}
\description{
This method first extracts the energy density, using the \code{\link{energyDensity}} method, and then
calculates onsets based on the energy density.
}

