% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyDensity.R
\name{plot.energyDensity}
\alias{plot.energyDensity}
\title{Plotting of EnergyDensity Objects}
\usage{
\method{plot}{energyDensity}(x, db = c(0, -1, -2, -3, -4, -5, -6, -10, -15,
  -Inf), limit = c(0.1, 0.01), limit.lty = 3, ...)
}
\arguments{
\item{x}{Energy density object to be plotted}

\item{db}{Decibel indicator positions in plot (set to \code{NULL} to deactivate)}

\item{limit}{Position of limit indicator (set to \code{NULL} to deactivate)}

\item{limit.lty}{Line type of limit indicator}

\item{...}{Additional arguments passed to \link{plot}}
}
\description{
Plotting of EnergyDensity Objects
}

