% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWave.R
\name{read.wav}
\alias{read.wav}
\title{Reads Wave Date From a File}
\usage{
read.wav(filename, channels = c("both", "left", "right"))
}
\arguments{
\item{filename}{The filename to be read}

\item{channels}{The channel(s) that should be used. The wave file is automatically
reduced to a mono wave file, either by selecting a single channel (values 
\code{left} or \code{right}) or by averaging accros both channels (\code{both}).
If the file is alread a mono file, this argument is ignored.}
}
\value{
An object of type \code{WaveData}
}
\description{
This function reads wave data from a file. The data is read using
the \code{\link[tuneR]{readWave}} function and then automatically
simplified for further processing.
}

