% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsHelpers.R
\name{slice}
\alias{slice}
\title{Generic function to partition a time series into windows with overlap}
\usage{
slice(x, window.width, stepsize, window.function = signal::hanning, ...)
}
\arguments{
\item{x}{Time series to be sliced}

\item{window.width}{Width of each partition window (in ms)}

\item{stepsize}{Stepsize for the windowing (in ms)}

\item{window.function}{Windowing function to be used during extraction (default: hanning window)}

\item{...}{Extra arguments for future methods}
}
\value{
A matrix with one window per column
}
\description{
Generic function to partition a time series into windows with overlap
}

