% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{spectrum}
\alias{spectrum}
\title{S3 Generic Function to Create a Spectrum From a Time Series}
\usage{
spectrum(x, window.width, stepsize, padding = TRUE,
  window.function = signal::hanning, ...)
}
\arguments{
\item{x}{Time Series for which the spectrum should be 
calculated}

\item{window.width}{Width of each spectrum arrea (in ms)}

\item{stepsize}{Stepsize for the windows (in ms)}

\item{padding}{If this is set to TRUE (default), the windows are
zero padded to the next power of two for efficiency}

\item{window.function}{Windowing function to be applied to the 
signal frames}

\item{...}{Passed to future methods.}
}
\value{
A matrix of the spectrum data with one time step in each column
}
\description{
S3 Generic Function to Create a Spectrum From a Time Series
}

