% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse.directory}
\alias{analyse.directory}
\title{Analyse a Complete Directory for Onset Times}
\usage{
analyse.directory(dirname, read.params = list(), filter = list(),
  onset.params = list(), energy.params = list(), quiet = TRUE)
}
\arguments{
\item{dirname}{Name of the directory from which the files should be taken}

\item{read.params}{Parameters passed to \code{\link{read.wav}} (except filter parameters)}

\item{filter}{List of filter parameters (set to \code{NULL} to deactivate filter). 
See "Details"}

\item{onset.params}{Parameters passed to \code{\link{onsets}}}

\item{energy.params}{Additional Parameters passed to \code{\link{energyDensity}}.}

\item{quiet}{If set to true, the function will output the current filename to be analysed}
}
\description{
Analyse a Complete Directory for Onset Times
}

