% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse.directory}
\alias{analyse.directory}
\title{Analyse a Complete Directory for Onset Times}
\usage{
analyse.directory(dirname, channels = c("both", "left", "right"),
  limit = 0.1, limit.type = c("absolute", "relative"),
  energy.params = list(), quiet = TRUE)
}
\arguments{
\item{dirname}{Name of the directory from which the files should be taken}

\item{channels}{The channel(s) that should be used. The wave file is automatically
reduced to a mono wave file, either by selecting a single channel (values 
\code{left} or \code{right}) or by averaging accros both channels (\code{both}).
If the file is alread a mono file, this argument is ignored.}

\item{limit}{Parameter used for filtering silence. Energy below this level
is ignored.}

\item{limit.type}{Type of limit, which is used to detect on- and offsets.
If set to \code{absolute} (default) then the limit value is used
directly. If set to \code{relative} then the limit is determined
as the energy in the i-th percentile, where i is given by the
limit parameter.}

\item{energy.params}{Additional Parameters passed to \code{\link{energyDensity}}.}

\item{quiet}{If set to true, the function will output the current filename to be analysed}
}
\description{
Analyse a Complete Directory for Onset Times
}

