% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse.directory}
\alias{analyse.directory}
\title{Analyse a Complete Directory for Onset Times}
\usage{
analyse.directory(dirname, channels = c("both", "left", "right"),
  limit = 0.1, window.width = 10, stepsize = 5, normalize = 0.9,
  window.function = signal::hanning)
}
\arguments{
\item{dirname}{Name of the directory from which the files should be taken}

\item{channels}{The channel(s) that should be used. The wave file is automatically
reduced to a mono wave file, either by selecting a single channel (values 
\code{left} or \code{right}) or by averaging accros both channels (\code{both}).
If the file is alread a mono file, this argument is ignored.}

\item{limit}{Parameter used for filtering silence. Energy below this level
(compared to all other energy levels in term of percentiles) is ignored.}

\item{window.width}{The windowing width that is used to estimate the density (in ms). The time
series data is chopped into windows of this length, and the energy density is calculated
for each of these windows seperately. Defaults to 10ms.}

\item{stepsize}{The stepsize to be used for windowing. Windows are advanced this many ms for each
calculation. Can be used to create overlapping windows. Defaults to 5ms.}

\item{normalize}{Normalization level for energy density values. Maximum energy density will be set to this
level to simplify comparison between data files. Set to 0 to deactivate normalization. 
Defaults to 0.9. Maximum normalization to 1.}

\item{window.function}{The windowing function to be used. This argument can be used to give
different weigth to different samples within the window. Several usable windowing functions
   are defined in the signal package. Defaults to \code{\link[signal]{boxcar}}, i.e. equal weight
   for all samples.}
}
\description{
Analyse a Complete Directory for Onset Times
}

