% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse.file}
\alias{analyse.file}
\title{Analyse a Single File for Onset Times}
\usage{
analyse.file(filename, read.params = list(), filter = list(),
  onset.params = list(), energy.params = list(), quiet = TRUE)
}
\arguments{
\item{filename}{Name of the file to be analyzed}

\item{read.params}{Parameters passed to \code{\link{read.wav}} (except filter parameters)}

\item{filter}{Filter parameters passed to \code{\link{read.wav}}}

\item{onset.params}{Parameters passed to \code{\link{onsets}}}

\item{energy.params}{Additional Parameters passed to \code{\link{energyDensity}}.}

\item{quiet}{If set to true, the function will output the current filename to be analysed}
}
\description{
Analyse a Single File for Onset Times
}

