% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse.file}
\alias{analyse.file}
\title{Analyse a Single File for Onset Times}
\usage{
analyse.file(filename, channels = c("both", "left", "right"), limit = 0.1,
  window.width = 10, stepsize = 5, window.function = signal::boxcar)
}
\arguments{
\item{filename}{The filename to be read}

\item{channels}{The channel(s) that should be used. The wave file is automatically
reduced to a mono wave file, either by selecting a single channel (values 
\code{left} or \code{right}) or by averaging accros both channels (\code{both}).
If the file is alread a mono file, this argument is ignored.}

\item{limit}{Parameter used for filtering silence. Energy below this level
(compared to all other energy levels in term of percentiles) is ignored.}

\item{window.width}{The windowing width that is used to estimate the density (in ms). The time
series data is chopped into windows of this length, and the energy density is calculated
for each of these windows seperately. Defaults to 10ms.}

\item{stepsize}{The stepsize to be used for windowing. Windows are advanced this many ms for each
calculation. Can be used to create overlapping windows. Defaults to 5ms.}

\item{window.function}{The windowing function to be used. This argument can be used to give
different weigth to different samples within the window. Several usable windowing functions
   are defined in the signal package. Defaults to \code{\link[signal]{boxcar}}, i.e. equal weight
   for all samples.}
}
\description{
Analyse a Single File for Onset Times
}

