% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvOutput.R
\name{expOnsets.as.csv2}
\alias{expOnsets.as.csv2}
\title{Analyse All Files from a Given Directory and Write Results as CSV file}
\usage{
expOnsets.as.csv2(dirname, filename, read.params = list(), filter = list(),
  onset.params = list(), energy.params = energy.params, ...)
}
\arguments{
\item{dirname}{Name of the directory from which the files should be taken}

\item{filename}{The filename of the csv file to which the results should be written}

\item{read.params}{Parameters passed to \code{\link{read.wav}} (except filter parameters)}

\item{filter}{List of filter parameters (set to \code{NULL} to deactivate filter). 
See "Details"}

\item{onset.params}{Parameters passed to \code{\link{onsets}}}

\item{energy.params}{Additional Parameters passed to \code{\link{energyDensity}}.}

\item{...}{Optional arguments passed to \code{\link[utils]{write.csv2}}}
}
\description{
This function uses the formating defined by \code{\link[utils]{write.csv2}}
}

