% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprinting.R
\name{fingerprint.ts}
\alias{fingerprint.ts}
\title{Extract Fingerprints for any Kind of ts Object}
\usage{
\method{fingerprint}{ts}(ts, feature.type = c("MFCCs", "MFCC", "spectrum"),
  feature.params = list(), ...)
}
\arguments{
\item{ts}{The time series object from which the fingerprint should
be extracted}

\item{feature.type}{The type of features to be used in fignerprinting}

\item{feature.params}{Additional arguments for the feature extractor}

\item{...}{Arguments passed to fingerprinting function for the features}
}
\description{
Fingerprints from ts objects are extracted by first extracting a feature
time series from the original time series and then running the appropriate
fingerprint extractor on this feature series.
}

