% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset.R
\name{onsets.energyDensity}
\alias{onsets.energyDensity}
\title{Extracts Onsets from an energyDensity Object}
\usage{
\method{onsets}{energyDensity}(ts, limit = 0.1, ...)
}
\arguments{
\item{ts}{The object containing the time series data.}

\item{limit}{Parameter used for filtering silence. Energy below this level
(compared to all other energy levels in term of percentiles) is ignored.}

\item{...}{ignored}
}
\description{
Extracts Onsets from an energyDensity Object
}

