% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset.R
\name{onsets.energyDensity}
\alias{onsets.energyDensity}
\title{Extracts Onsets from an energyDensity Object}
\usage{
\method{onsets}{energyDensity}(ts, limit = 0.1, limit.type = c("absolute",
  "relative"), ...)
}
\arguments{
\item{ts}{The object containing the time series data.}

\item{limit}{Parameter used for filtering silence. Energy below this level
is ignored.}

\item{limit.type}{Type of limit, which is used to detect on- and offsets.
If set to \code{absolute} (default) then the limit value is used
directly. If set to \code{relative} then the limit is determined
as the energy in the i-th percentile, where i is given by the
limit parameter.}

\item{...}{ignored}
}
\description{
Extracts Onsets from an energyDensity Object
}

