% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset.R
\name{onsets.WaveData}
\alias{onsets.WaveData}
\title{Specialized Method for Extracting Onsets from WaveData objects}
\usage{
\method{onsets}{WaveData}(ts, limit = 0.1, window.width = 10,
  stepsize = 5, window.function = signal::boxcar, ...)
}
\arguments{
\item{ts}{The object containing the time series data.}

\item{limit}{Parameter used for filtering silence. Energy below this level
(compared to all other energy levels in term of percentiles) is ignored.}

\item{window.width}{The windowing width that is used to estimate the density (in ms). The time
series data is chopped into windows of this length, and the energy density is calculated
for each of these windows seperately. Defaults to 10ms.}

\item{stepsize}{The stepsize to be used for windowing. Windows are advanced this many ms for each
calculation. Can be used to create overlapping windows. Defaults to 5ms.}

\item{window.function}{The windowing function to be used. This argument can be used to give
different weigth to different samples within the window. Several usable windowing functions
   are defined in the signal package. Defaults to \code{\link[signal]{boxcar}}, i.e. equal weight
   for all samples.}

\item{...}{ignored}
}
\description{
This method first extracts the energy density, using the \code{\link{energyDensity}} method, and then
calculates onsets based on the energy density.
}

